+++ src/schemathesis/schemas.py
@@ -49,7 +49,6 @@ from .utils import PARAMETRIZE_MARKER, GivenInput, given_proxy
 if TYPE_CHECKING:
     import hypothesis
     from hypothesis.strategies import SearchStrategy
-    from hypothesis.vendor.pretty import RepresentationPrinter
     from pyrate_limiter import Limiter
 
     from .stateful import Stateful, StatefulTest
@@ -103,8 +102,7 @@ class BaseSchema(Mapping):
     def __post_init__(self) -> None:
         self.hook = to_filterable_hook(self.hooks)  # type: ignore[method-assign]
 
-    def _repr_pretty_(self, printer: RepresentationPrinter, cycle: bool) -> None:
-        return None
+    def _repr_pretty_(self, *args: Any, **kwargs: Any) -> None: ...
 
     def include(
         self,