+++ src/schemathesis/transports/__init__.py
@@ -21,7 +21,6 @@ if TYPE_CHECKING:
     import requests
     import werkzeug
     from _typeshed.wsgi import WSGIApplication
-    from hypothesis.vendor.pretty import RepresentationPrinter
     from starlette_testclient._testclient import ASGI2App, ASGI3App
 
     from ..models import Case
@@ -35,8 +34,7 @@ class RequestConfig:
     proxy: str | None = None
     cert: RequestCert | None = None
 
-    def _repr_pretty_(self, printer: RepresentationPrinter, cycle: bool) -> None:
-        return None
+    def _repr_pretty_(self, *args: Any, **kwargs: Any) -> None: ...
 
     @property
     def prepared_timeout(self) -> float | None: